+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- onstart handler, that allows to add code that shall be only executed before the actual workflow execution (not on dryrun).
+- Parameters defined in the cluster config file are now accessible in the job properties under the key "cluster".
 ### Changed
 - Allow to use rule/job parameters with braces notation in cluster config.
+- Show a proper error message in case of recursion errors.
+- Remove non-empty temp dirs.
+- Don't set the process group of Snakemake in order to allow kill signals from parent processes to be propagated.
 
 ## [3.5.5] - 2016-01-23
 ### Added